SalCodec/codec/requirements.txt
imageio==2.21.2
jsonschema==4.16.0
jupyter==1.0.0
matplotlib==3.3.4
notebook==6.4.12
numpy==1.23.2
opencv-python==4.6.0.66
pillow==9.2.0
protobuf==3.19.4
pytorch-msssim==0.2.0
scikit-image==0.19.3
tensorboard==2.10.0
tensorboardx==2.5.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0